chore(planets.model): import mongoose model
Import mongoose planets model and add create method to add the planets to mongo db. Since the code is implemented in the loadPlanetsData() function, which is call in the server, it is best to comment out the the planets.create method so that everytime the function is called in a different instance of the server that is a part of another node js process in the same cluster, the data is added repeatedly to the collection, meaning that multiple copies of the same document will be added to the same collection. The planets.upsert method will be used in the future instead of the create method.