refactor(launches.mongo): change schema
Take out the required property from the target object that is  apart of the launches schema. This was done because the data requested from the spaceX api has its own planets (or targeted planets) in it launches history. This makes sure that the planets in the database do not have any conflict with the planets set as a response from the spaceX api.